+++ packages/contracts/contracts/test/TestSharedPlugin.sol
@@ -2,15 +2,16 @@
 
 pragma solidity 0.8.10;
 
-import "@openzeppelin/contracts-upgradeable/proxy/utils/Initializable.sol";
+import {Initializable} from "@openzeppelin/contracts-upgradeable/proxy/utils/Initializable.sol";
 import {IPermissionOracle} from "../core/permission/IPermissionOracle.sol";
 
-import "../core/component/Component.sol";
+import {PluginUUPSUpgradeable} from "../core/plugin/PluginUUPSUpgradeable.sol";
 import {DaoUnauthorized} from "../utils/auth.sol";
+import {IDAO} from "../core/IDAO.sol";
 
-/// @notice A test component that manages permission to internal objects by associating their IDs with specific DAOs. Only the DAO for which the object was created has the permission to perform ID-gated actions on them.
+/// @notice A test Plugin that manages permission to internal objects by associating their IDs with specific DAOs. Only the DAO for which the object was created has the permission to perform ID-gated actions on them.
 /// @dev This is realized by asking an `IPermissionOracle` that must be authorized in the DAO's permission manager.
-contract TestSharedComponent is Component {
+contract TestSharedPlugin is PluginUUPSUpgradeable {
     bytes32 public constant ID_GATED_ACTION_PERMISSION_ID = keccak256("ID_GATED_ACTION_PERMISSION");
 
     mapping(uint256 => IDAO) public ownedIds;
@@ -38,7 +39,7 @@ contract TestSharedComponent is Component {
     }
 
     function initialize(IDAO _dao) external initializer {
-        __Component_init(_dao);
+        __PluginUUPSUpgradeable_init(_dao);
     }
 
     /// @notice Creates a new object with an ID being associated with a specific DAO
@@ -58,7 +59,7 @@ contract TestSharedComponent is Component {
     }
 }
 
-/// @notice The oracle associated with `TestSharedComponent`
+/// @notice The oracle associated with `TestSharedPlugin`
 contract TestIdGatingOracle is IPermissionOracle {
     uint256 public allowedId;