+++ packages/contracts/test/core/permission/integration-tests/shared-plugin.ts
@@ -3,7 +3,7 @@ import {ethers} from 'hardhat';
 import {SignerWithAddress} from '@nomiclabs/hardhat-ethers/signers';
 
 import {
-  TestSharedComponent,
+  TestSharedPlugin,
   TestIdGatingOracle,
   DAO,
 } from '../../../../typechain';
@@ -13,9 +13,9 @@ const ID_GATED_ACTION_PERMISSION_ID = ethers.utils.id(
   'ID_GATED_ACTION_PERMISSION'
 );
 
-describe('SharedComponent', function () {
+describe('SharedPlugin', function () {
   let signers: SignerWithAddress[];
-  let testComponent: TestSharedComponent;
+  let testPlugin: TestSharedPlugin;
   let managingDAO: DAO;
   let dao1: DAO;
   let dao2: DAO;
@@ -39,18 +39,18 @@ describe('SharedComponent', function () {
     await dao1.initialize('0x', ownerAddress, ethers.constants.AddressZero);
     await dao2.initialize('0x', ownerAddress, ethers.constants.AddressZero);
 
-    // Deploy the `TestSharedComponent`
-    const TestSharedComponent = await ethers.getContractFactory(
-      'TestSharedComponent'
+    // Deploy the `TestSharedPlugin`
+    const TestSharedPlugin = await ethers.getContractFactory(
+      'TestSharedPlugin'
     );
-    testComponent = await TestSharedComponent.deploy();
-    await testComponent.initialize(managingDAO.address);
+    testPlugin = await TestSharedPlugin.deploy();
+    await testPlugin.initialize(managingDAO.address);
 
     expectedUnauthorizedError = customError(
       'DaoUnauthorized',
       managingDAO.address,
-      testComponent.address,
-      testComponent.address,
+      testPlugin.address,
+      testPlugin.address,
       ownerAddress,
       ID_GATED_ACTION_PERMISSION_ID
     );
@@ -58,15 +58,15 @@ describe('SharedComponent', function () {
 
   it('increments IDs', async () => {
     expect(
-      await testComponent.callStatic.createNewObject(dao1.address)
+      await testPlugin.callStatic.createNewObject(dao1.address)
     ).to.be.equal(0);
 
-    const tx = await testComponent.createNewObject(dao1.address);
+    const tx = await testPlugin.createNewObject(dao1.address);
     await tx.wait();
     await ethers.provider.send('evm_mine', []);
 
     expect(
-      await testComponent.callStatic.createNewObject(dao1.address)
+      await testPlugin.callStatic.createNewObject(dao1.address)
     ).to.be.equal(1);
   });
 
@@ -83,19 +83,19 @@ describe('SharedComponent', function () {
 
       // Grants signers[0] the permission to do ID gated actions with the deployed `TestIdGatingOracle` oracle
       dao1.grantWithOracle(
-        testComponent.address,
+        testPlugin.address,
         ownerAddress,
         ID_GATED_ACTION_PERMISSION_ID,
         oracle.address
       );
 
-      // Deploy a new object in the `TestComponent` which will have the ID 0
-      const tx = await testComponent.createNewObject(dao1.address);
+      // Deploy a new object in the `TestPlugin` which will have the ID 0
+      const tx = await testPlugin.createNewObject(dao1.address);
       await tx.wait();
       await ethers.provider.send('evm_mine', []);
 
       // Check that the ID gated action can be executed
-      await expect(testComponent.callStatic.idGatedAction(allowedId)).to.not.be
+      await expect(testPlugin.callStatic.idGatedAction(allowedId)).to.not.be
         .reverted;
     });
 
@@ -109,7 +109,7 @@ describe('SharedComponent', function () {
 
       // Grants signers[0] the permission to do ID gated actions with the deployed `TestIdGatingOracle` oracle
       dao1.grantWithOracle(
-        testComponent.address,
+        testPlugin.address,
         ownerAddress,
         ID_GATED_ACTION_PERMISSION_ID,
         oracle.address
@@ -117,21 +117,21 @@ describe('SharedComponent', function () {
 
       // The call fails because no object with ID 1 exists
       await expect(
-        testComponent.callStatic.idGatedAction(nonExistingId)
+        testPlugin.callStatic.idGatedAction(nonExistingId)
       ).to.be.revertedWith(customError('ObjectIdNotAssigned', nonExistingId));
 
       // Create object with ID 0
-      let tx = await testComponent.createNewObject(dao1.address);
+      let tx = await testPlugin.createNewObject(dao1.address);
       await tx.wait();
       await ethers.provider.send('evm_mine', []);
 
       // The call still fails because no object with ID 1 exists
       await expect(
-        testComponent.callStatic.idGatedAction(nonExistingId)
+        testPlugin.callStatic.idGatedAction(nonExistingId)
       ).to.be.revertedWith(customError('ObjectIdNotAssigned', nonExistingId));
 
       // The call executes for the allowed ID 0
-      await expect(testComponent.callStatic.idGatedAction(allowedId)).to.not.be
+      await expect(testPlugin.callStatic.idGatedAction(allowedId)).to.not.be
         .reverted;
     });
 
@@ -143,35 +143,35 @@ describe('SharedComponent', function () {
       const Oracle = await ethers.getContractFactory('TestIdGatingOracle');
       oracle = await Oracle.deploy(allowedId);
 
-      // Grants signers[0] the permission to do ID gated actions on `testComponent` via `oracle`
+      // Grants signers[0] the permission to do ID gated actions on `testPlugin` via `oracle`
       dao1.grantWithOracle(
-        testComponent.address,
+        testPlugin.address,
         ownerAddress,
         ID_GATED_ACTION_PERMISSION_ID,
         oracle.address
       );
       dao2.grantWithOracle(
-        testComponent.address,
+        testPlugin.address,
         ownerAddress,
         ID_GATED_ACTION_PERMISSION_ID,
         oracle.address
       );
 
       // Create ID-gated object associated with `dao1`
-      let tx = await testComponent.createNewObject(dao1.address);
+      let tx = await testPlugin.createNewObject(dao1.address);
       await tx.wait();
-      tx = await testComponent.createNewObject(dao2.address);
+      tx = await testPlugin.createNewObject(dao2.address);
       await tx.wait();
 
       await ethers.provider.send('evm_mine', []);
 
       // The call is allowed for the allowed ID
-      await expect(testComponent.callStatic.idGatedAction(allowedId)).to.not.be
+      await expect(testPlugin.callStatic.idGatedAction(allowedId)).to.not.be
         .reverted;
 
       // The call fails if the ID differs
       await expect(
-        testComponent.callStatic.idGatedAction(existingButNotAllowedId)
+        testPlugin.callStatic.idGatedAction(existingButNotAllowedId)
       ).to.be.revertedWith(expectedUnauthorizedError);
     });
 
@@ -183,12 +183,12 @@ describe('SharedComponent', function () {
       oracle = await Oracle.deploy(allowedId);
 
       // Create ID-gated object associated with `dao1`
-      const tx = await testComponent.createNewObject(dao1.address);
+      const tx = await testPlugin.createNewObject(dao1.address);
       await tx.wait();
       await ethers.provider.send('evm_mine', []);
 
       await expect(
-        testComponent.callStatic.idGatedAction(allowedId)
+        testPlugin.callStatic.idGatedAction(allowedId)
       ).to.be.revertedWith(expectedUnauthorizedError);
     });
 
@@ -201,19 +201,19 @@ describe('SharedComponent', function () {
 
       // Grants signers[0] the permission to do ID gated actions with the deployed `TestIdGatingOracle` oracle
       dao2.grantWithOracle(
-        testComponent.address,
+        testPlugin.address,
         ownerAddress,
         ID_GATED_ACTION_PERMISSION_ID,
         oracle.address
       );
 
       // Create ID-gated object associated with `dao1`
-      const tx = await testComponent.createNewObject(dao1.address);
+      const tx = await testPlugin.createNewObject(dao1.address);
       await tx.wait();
       await ethers.provider.send('evm_mine', []);
 
       await expect(
-        testComponent.callStatic.idGatedAction(allowedId)
+        testPlugin.callStatic.idGatedAction(allowedId)
       ).to.be.revertedWith(expectedUnauthorizedError);
     });
 
@@ -226,19 +226,19 @@ describe('SharedComponent', function () {
 
       // Grants signers[0] the permission to do ID gated actions with the deployed `TestIdGatingOracle` oracle
       dao1.grantWithOracle(
-        testComponent.address,
+        testPlugin.address,
         ownerAddress,
         ID_GATED_ACTION_PERMISSION_ID,
         oracle.address
       );
 
       // Create ID-gated object associated with `dao1`
-      const tx = await testComponent.createNewObject(dao2.address);
+      const tx = await testPlugin.createNewObject(dao2.address);
       await tx.wait();
       await ethers.provider.send('evm_mine', []);
 
       await expect(
-        testComponent.callStatic.idGatedAction(allowedId)
+        testPlugin.callStatic.idGatedAction(allowedId)
       ).to.be.revertedWith(expectedUnauthorizedError);
     });
   });